+++ host-frontend-root/frontend-src-root/src/application-business-rules/ports/gateway/IRewriteRuleRepository.ts
