+++ host-frontend-root/frontend-src-root/src/application/usecases/rule/LoadRewriteRuleForEditUseCase.ts
@@ -1,4 +1,4 @@
-import { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
+import { IRewriteRuleRepository } from 'src/application-business-rules/ports/gateway/IRewriteRuleRepository';
 import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 export class LoadRewriteRuleForEditUseCase {
