+++ host-frontend-root/frontend-src-root/tests/unit/application-business-rules/interactors/ToggleRuleActiveInteractor/mocks/createMockRepository.ts
@@ -1,6 +1,6 @@
 import { vi } from 'vitest';
 
-import { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
+import { IRewriteRuleRepository } from 'src/application-business-rules/ports/gateway/IRewriteRuleRepository';
 
 /**
  * IRewriteRuleRepositoryのモックオブジェクトを生成する
